refactor: split dispatcher\models.py into 2 files
